backend/requirements.txt
huggingface_hub
dotenv
sqlalchemy
pydantic
fastapi
uvicorn
psycopg2-binary
passlib
argon2_cffi
python-jose
pytest